+++ public/views/home.jade
@@ -1,0 +1,6 @@
+extends layout
+
+block content
+  h1= title
+  p Welcome to #{title}
+  a(href="/login") Log In
