+++ public/views/layout.jade
@@ -1,0 +1,11 @@
+doctype html
+html
+  head
+    title= title
+    link(rel='stylesheet', href='/css/style.css')
+    script(src="http://ajax.googleapis.com/ajax/libs/angularjs/1.3.10/angular.min.js")
+    script(src="http://cdnjs.cloudflare.com/ajax/libs/angular-ui-router/0.2.10/angular-ui-router.js")
+    script(src="/javascripts/app.js")
+  body(ng-app="timeApp")
+    block content
+
