--- public/views/login.html
@@ -1,23 +1,0 @@
-<form>
-  <h1>User Log in</h1>
-  <div class="inset">
-  <p>
-    <label for="email">EMAIL ADDRESS</label>
-    <input type="text" name="email" id="email">
-  </p>
-  <p>
-    <label for="password">PASSWORD</label>
-    <input type="password" name="password" id="password">
-  </p>
-  
-  <p>
-    <input type="checkbox" name="remember" id="remember">
-    <label for="remember">Remember me</label>
-  </p>
-  </div>
-  <p class="p-container">
-    <span>Forgot password ?</span>
-    <input type="submit" name="go" id="go" value="Log in">
-  </p>
-</form>
-
