+++ server.js
@@ -17,7 +17,7 @@ var session = require('express-session')
 
 // route files
 var routes = require('./routes/index');
-var users = require('./routes/users');
+var auth = require('./routes/auth');
 var api = require('./routes/api');
 
 var app = express();
@@ -48,7 +48,7 @@ app.use(cookieParser());
 
 
 app.use('/', routes);
-//app.use('/users', users);
+app.use('/auth', auth);
 app.use('/api', api);
